requirements.txt
requests
icalendar
datetime
pyyaml
matplotlib
